backend/requirements.txt
# Django
django
django-filter
django-cors-headers

# DRF
djangorestframework
djangorestframework-camel-case
drf-flex-fields

# Other
python-dateutil
